code/slave/raspi/requirements.txt
decorator==4.0.10
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pygame==1.9.2b8 --allow-external pygame --allow-unverified pygame
Pygments==2.1.3
pyserial==3.2.1
pyzmq==16.0.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7


ansible-playbook -vvv -i "localhost," -c local /root/Zoomachine-ansible/installation/playbook.yml
